+++ src/main/java/main/TaxYear2017.java
@@ -6,19 +6,21 @@ import entity.TaxPayer;
 
 import javax.ejb.Stateful;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Created by SamkeDl on 12/07/2017.
  */
 @Stateful
 public class TaxYear2017 implements Serializable {
+    /*
     public static TaxPayer calculateTax(Person person,String incomeFrequency,int medicalAidDeduction){
         TaxPayer taxPayer = new TaxPayer();
         taxPayer.setIncome(person.getIncome());
         taxPayer.setTaxableIncome(person.getIncome());
 
         //tax before credits
-        double taxBeforeCredits = getTaxRate(person.getIncome());
+        BigDecimal taxBeforeCredits = getTaxRate(person.getIncome());
         taxPayer.setAnnualTaxBefore(taxBeforeCredits);
         taxPayer.setMonthlyTaxBefore(taxBeforeCredits/12);
         taxPayer.setTaxBeforeCredits(taxBeforeCredits);
@@ -62,9 +64,9 @@ public class TaxYear2017 implements Serializable {
 
         return taxPayer;
     }
-    private static double getTaxRate(int income){
+    private static BigDecimal getTaxRate(BigDecimal income){
         System.out.println("Calculating tax on "+income+" income");
-        double tax = 0;
+        BigDecimal tax = 0;
         if (income <= 188000) {
             tax = income * 0.18;
             System.out.println("TAX calculated on 18%");
@@ -119,5 +121,5 @@ public class TaxYear2017 implements Serializable {
             return  TERTIARY;
         return 0;
     }
-
+*/
 }
